prk final/requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
pandas
matplotlib
